Refactor vyos.vyos jobs
With this change, we are now making it more important to use unit
testing to check for python versions over integration jobs.  This is
mostly to help cut down on the amount of time it takes a PR to run.

The idea is, integration testing hasn't been a good indicator of finding
issues with the different versions of python, and we likely can improve
unitesting to account for that.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>